chore: extract theme providers
